# docs/swagger.yaml
basePath: /api
definitions:
  agenda.CreateAgendaRequest:
    properties:
      creator_id:
        type: string
      description:
        type: string
      room_id:
        type: string
      scheduled_at:
        type: string
      title:
        type: string
    required:
    - creator_id
    - room_id
    - scheduled_at
    - title
    type: object
  auth.LoginResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/entities.User'
    type: object
  chat.SendMessageRequest:
    properties:
      content:
        type: string
      room_id:
        type: string
      sender_id:
        type: string
      type:
        type: string
    required:
    - content
    - room_id
    - sender_id
    type: object
  entities.Agenda:
    properties:
      created_at:
        type: string
      creator:
        $ref: '#/definitions/entities.User'
      creator_id:
        type: string
      description:
        type: string
      id:
        type: string
      room_id:
        type: string
      scheduled_at:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  entities.Message:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      room_id:
        type: string
      sender:
        $ref: '#/definitions/entities.User'
      sender_id:
        type: string
      type:
        description: TEXT, IMAGE, FILE
        type: string
      updated_at:
        type: string
    type: object
  entities.Room:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/entities.User'
        type: array
      type:
        description: PRIVATE, GROUP
        type: string
      updated_at:
        type: string
    type: object
  entities.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: string
      last_seen:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  handlers.GoogleLoginRequest:
    properties:
      idToken:
        type: string
    required:
    - idToken
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  room.CreateRoomRequest:
    properties:
      name:
        type: string
      participants:
        items:
          type: string
        type: array
      type:
        description: '"private" or "group"'
        type: string
    required:
    - participants
    - type
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample chat server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Chat Golang API
  version: "1.0"
paths:
  /auth/google:
    post:
      consumes:
      - application/json
      description: Authenticate user using Google ID Token
      parameters:
      - description: Google Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Google Login
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get current user profile using JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with username, email, and password
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - auth
  /rooms:
    get:
      consumes:
      - application/json
      description: Get all chat rooms that the current user belongs to
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Room'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get my rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Create a private or group chat room
      parameters:
      - description: Create Room Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/room.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Room'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - rooms
  /rooms/{roomId}/agendas:
    get:
      consumes:
      - application/json
      description: Get all agendas for a specific room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Agenda'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get agendas for a room
      tags:
      - agenda
    post:
      consumes:
      - application/json
      description: Create a new agenda for a room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Create Agenda Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/agenda.CreateAgendaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new agenda
      tags:
      - agenda
  /rooms/{roomId}/messages:
    get:
      consumes:
      - application/json
      description: Get message history for a specific room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Message'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get messages for a room
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Send a message to a room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Send Message Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Message'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Send a message
      tags:
      - chat
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.User'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific user by their UUID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.User'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT Anda. Contoh: "eyJhbGci..." atau "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"